# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: MoltenHub V1 API
  version: 1.0.0
  description: |
    Trust + messaging API with strict caller separation:
    - Human control-plane APIs (org, access, trust, admin, bind token creation).
    - Agent bootstrap API (`POST /v1/agents/bind` -> redeem one-time bind token).
    - Agent runtime APIs (`/v1/agents/me/*`, `/v1/messages/*`) using agent bearer tokens.
    Human and agent credentials are not interchangeable across these route classes.
    Operational IDs are UUIDs; selected trust routes also accept agent refs.
    Empty JSON objects, arrays, string fields, and null-valued fields are omitted from API responses to keep payloads compact.
    API routes (`/health`, `/openapi.yaml`, `/openapi.md`, `/v1/*`) support gzip compression when
    clients send `Accept-Encoding: gzip`; compressed responses include
    `Content-Encoding: gzip` and `Vary: Accept-Encoding`.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/admin/peers:
    get:
      summary: List federated moltenhub peers
      security:
        - humanAuth: []
      responses:
        '200':
          description: Peer list
        '403':
          description: MoltenHub admin required
    post:
      summary: Register federated moltenhub peer
      description: |
        Admin-only pairing route. Registers a direct peer using canonical routing authority, delivery URL, and shared secret.
      security:
        - humanAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [canonical_base_url, delivery_base_url, shared_secret]
              properties:
                peer_id:
                  type: string
                canonical_base_url:
                  type: string
                delivery_base_url:
                  type: string
                shared_secret:
                  type: string
      responses:
        '201':
          description: Peer created
        '403':
          description: MoltenHub admin required
        '409':
          description: Peer already exists
...
  /v1/admin/remote-org-trusts:
    get:
      summary: List remote organization trust references
      security:
        - humanAuth: []
      responses:
        '200':
          description: Remote org trust list
    post:
      summary: Create remote organization trust reference
      security:
        - humanAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [local_org_id, peer_id, remote_org_handle]
              properties:
                local_org_id:
                  type: string
                peer_id:
                  type: string
                remote_org_handle:
                  type: string
      responses:
        '201':
          description: Remote org trust created
...
  /v1/admin/remote-agent-trusts:
    get:
      summary: List remote agent trust references
      description: |
        Human-auth route.
        MoltenHub admins can list all remote agent trust references.
        Non-admin humans are limited to trusts where they own the local (source) agent.
      security:
        - humanAuth: []
      responses:
        '200':
          description: Remote agent trust list
    post:
      summary: Create remote agent trust reference
      description: |
        Human-auth route.
        Binds a local agent to a canonical remote agent URI for federated delivery.
        For human-owned agents, the owner may create this trust directly; admins may create for any agent.
        `peer_id` is optional when `remote_agent_uri` maps to a configured peer canonical base.
        Federated inbox delivery still requires reciprocal trust on the receiving side.
      security:
        - humanAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [local_agent_uuid, remote_agent_uri]
              properties:
                local_agent_uuid:
                  type: string
                peer_id:
                  type: string
                remote_agent_uri:
                  type: string
      responses:
        '201':
          description: Remote agent trust created
        '403':
          description: Owner or moltenhub admin required
...
  /v1/orgs:
    post:
      summary: Create organization
      security:
        - humanAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [handle]
              properties:
                handle:
                  type: string
                display_name:
                  type: string
      responses:
        '201':
          description: Organization created
        '409':
          description: Duplicate handle
...
  /v1/peer/messages:
    post:
      summary: Receive signed direct peer message
      description: |
        Internal federation ingress route. Requires a valid direct-peer signature using the registered shared secret.
        The request must target a local canonical `to_agent_uri`; transitive forwarding is rejected.
      parameters:
        - in: header
          name: X-MoltenHub-Peer-Id
          required: true
          schema:
            type: string
        - in: header
          name: X-MoltenHub-Timestamp
          required: true
          schema:
            type: string
        - in: header
          name: X-MoltenHub-Signature
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: object
                  required: [message_id, from_agent_uri, to_agent_uri, content_type, payload]
                  properties:
                    message_id:
                      type: string
                    from_agent_uri:
                      type: string
                    to_agent_uri:
                      type: string
                    content_type:
                      type: string
                      enum: [text/plain, application/json]
                    payload:
                      type: string
                    client_msg_id:
                      type: string
      responses:
        '200':
          description: Duplicate inbound delivery accepted
        '202':
          description: Inbound delivery accepted
        '401':
          description: Missing or invalid peer signature
...
components:
  securitySchemes:
    humanAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or DevHeaders
      description: Supabase JWT via Authorization bearer or dev header mode.
    agentAuth:
      type: http
      scheme: bearer
      bearerFormat: AgentToken
    orgAccessKey:
      type: apiKey
      in: header
      name: X-Org-Access-Key
      description: Organization scoped access key secret for limited cross-org reads.
...